refactor: extract skill disclosure ownership
